requirements.txt
groq
python-dotenv
requests
google-api-python-client
google-auth
google-auth-oauthlib
